grpc接口迁移完毕

+++ interface/conf.go
@@ -23,7 +23,7 @@ type Conf interface {
 
 	// SetConfig
 	// 设置一个配置变量（若存在则覆盖, 不存在不会创建，需要先注册）
-	SetConfParam(string, string, string, int)
+	SetConfParam(string, string, string, int64)
 
 	// ReloadConfig
 	// 重载配置
@@ -31,5 +31,5 @@ type Conf interface {
 
 	// RegisterConfParam
 	// 注册配置，已拥有，则不覆盖
-	RegisterConfParam(string, string, string, int, bool)
+	RegisterConfParam(string, string, string, int64, bool)
 }
